+++ packages/autocomplete-core/src/utils/index.ts
@@ -3,6 +3,8 @@ export * from './createCancelablePromiseList';
 export * from './createConcurrentSafePromise';
 export * from './getNextActiveItemId';
 export * from './getNormalizedSources';
+export * from './isChrome';
 export * from './getActiveItem';
 export * from './isOrContainsNode';
+export * from './isSamsung';
 export * from './mapToAlgoliaResponse';